apic: Fix apic unittest for latest qemu-kvm
Since commit b334ec56 in qemu-kvm the use of irqlines > 15
is not supported anymore. This causes the apic unittest to
fail since this commit. Since the commit-msg explicitly
states that potential users of this needs to be fixed, here
is the fix which uses irq lines 14 and 15 instead in the
unittest.

Cc: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Joerg Roedel <joerg.roedel@amd.com>
Reviewed-by: Amos Kong <akong@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>