Update unittests.cfg with list of vmexit subtests
vmexit takes a parameter to run appropriate subtests, but
unittests.cfg was not updated to reflect that. As this
config file is used by automated test harnesses such as
autotest, this will make the appropriate tests not
executed.

Tested successfuly with qemu-kvm shipped in Fedora 17
and RHEL 6.3. I've found problems with qemu-kvm compiled
from master HEAD, but it's not related to this patch, as
all other unittests are failing (currently under investigation).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>