x86: add INIT test
Test various ways of sending an INIT interrupt to the bootstrap processor.
In all cases, the processor should start executing from 0xfffffff0 and
other devices should be untouched.

With the patches I just sent to qemu-devel (also at branch x86-soft-reset
of git://github.com/bonzini/qemu.git), this passes with TCG and the
userspace irqchip.  It fails with in-kernel irqchip.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>