x86: api: allocate TSS region dynamically
Holes across the 3GB mark can occur on 64-bit machines.  Use posix_memalign
to reserve some space in the process's address space for the TSS, and
carve a second hole for it.  The memory itself is not used---it's only
that we want the process to not use that area while in guest context,
and the easiest way to do it is to allocate memory.

Allocations still are not allowed to go up to 4GB, because of the APIC
access page and EPT identity pagetable lie.  Detect the situation and
print a nice error message.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>