report: add prefix support
Add some methods to report in order to manage output prefixes.
Also add strstr to string, as it was useful for implementing
report_prefix_pop. Prefixes can be useful, as test code
frequently has the following pattern

main()
{
  foreach(subtest)
    run(subtest)
}

But, if we want output lines to include 'subtest: ', then we
need

  report("%s: ...", cond, subtest, ...)

for each report line. With push/pop we can now just do

  subtest()
  {
     report_prefix_push(subtest);

     report(...);
     ...
     report(...);

     report_prefix_pop();
  }

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>