arm/arm64: drop mmu_set_enabled
The mmu is enabled automatically for all cpus, they must disable it
themselves if they don't want it on. Switch from managing a cpumask
of enabled cpus to one of disabled cpus. This allows us to remove
the mmu_set_enabled call from secondary_cinit, and the function all
together.

Signed-off-by: Andrew Jones <drjones@redhat.com>
[Place CPUs in MMU disabled state at startup. - Paolo]
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>