mkstandalone: fix generation for arm
The mkstandalone rewrite now packages the architecture run scripts.
This brings many benefits, but at the cost of no longer being able
to source additional files from the run script, at least not while
in standalone mode. We can live with that, but arm/run was sourcing
config.mak, so we need to stop that in standalone mode, as well as
to provide the standalone arm/run script what it needs.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Message-Id: <1453474709-10679-3-git-send-email-drjones@redhat.com>
Reviewed-by: Radim Krčmář <rkrcmar@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>