x86: emulator: disable test_lldt
When compiled with -fno-omit-frame-pointer, test_lldt causes a triple
fault, which insta-reboots the guest. I couldn't figure out why the
triple-fault was happening, so I disabled the test outright. Knowing
that the other emulator.c tests pass is more useful than a VM stuck in
a reboot loop.

On x86-64, this test was adding little value since it was broken to
begin with: an LDT descriptor is 16 bytes on x86-64. The GDT limit
was also bogus. I've fixed these problems for posterity.

N.B. I only tested this on Linux 4.3.5. Perhaps lldt emulation has
been fixed since then.

Signed-off-by: Peter Feiner <pfeiner@google.com>
Message-Id: <1456967378-6367-3-git-send-email-pfeiner@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>