powerpc/spapr_hcall: Fix assembler constraints of the h_random h-call
As noted by Laurent Vivier, a register that is marked with "+r" as
input and output is not required to be listed in the set of input
registers anymore.
And as noted by David Gibson, the sPAPR hypercall is also allowed to
clobber a bunch of other registers. According to the LoPAPR spec,
r0 and r3-r12, CTR, XER and CC registers are volatile, so we should
mark them in the clobber list.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <1457004188-17186-1-git-send-email-thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>