report: move blank line to better side of SUMMARY
Currently test.log gets output of the form

 test1-command-line
 test1-output

 SUMMARY:
 test2-command-line
 test2-output

 SUMMARY:
 test3-command-line
 ...

Having a blank line before SUMMARY, and not after, it makes it
tough to group lines by eye. Change it to

 test1-command-line
 test1-output
 SUMMARY:

 test2-command-line
 test2-output
 SUMMARY:

 test3-command-line
 ...

The empty line is moved from report_summary to runtime in order
to remove it altogether when running tests individually or as a
standalone.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>