arch-run: fix terminal issues without stty
This patch reverts the stty parts of 8727c886 "runtime: better
handling of QEMU aborts", as we can more robustly fix the same
issue by simply redirecting /dev/null into qemu's stdin. The
obvious side effect of this is that serial input will no longer
be wired up to the tty (or any other input file descriptor), but
unit tests do not, and should not, be reading from the serial port
anyway.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>