Makefile: add explicit directories target
This is slightly clumsy way to ensure the directory structure for the
out-of-tree build is made first. The PHONY target is depended on by
make all first which runs mkdir for all OBJDIRS.

If an architecture target needs deeper nesting it needs to add
directories to OBJDIRS so it can be picked up by the directories rule.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>