Makefiles: fix up the x86 build dirs and include/link paths
We still need to tell the compiler the correct search path for finding
headers and the like. This is slightly complicated by the "dynamic"
asm search path which is in our build tree but (may be) symlinked to
the right architectures headers.

Also we explicitly include SRCDIR for the linking scripts as VPATH
doesn't seem to find them well enough.

Also update OBJDIRS to include lib/x86.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>