hyperv: don't take vcpu where inappropriate
MSRs can not be written on a forein cpu, so the functions to create /
destroy sint routes, which also enable / disable the respective SINTs,
can only run on the current cpu.  To avoid confusion, stop taking vcpu
argument and use smp_id() internally.

While chaning the function signatures, also adjust them to be explicit
about the width of input values.

Signed-off-by: Roman Kagan <rkagan@virtuozzo.com>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>