s390x: Fix bad dependency on asm-offsets.h
When building the kvm-unit-tests for s390x in parallel ("make -j10"
for example), the build sometimes fails since asm-offsets.h has not
been generated yet. The problem is a bad rule in the Makefile: The
test cases are stored in the the "tests" variable, and not in
"test_cases" (which is just an abstract target, not a variable).

Signed-off-by: Thomas Huth <thuth@redhat.com>
Reviewed-by: David Hildenbrand <david@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>